requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
redis==5.2.0

# Audio processing
librosa==0.10.2
soundfile==0.12.1
pydub==0.25.1
ffmpeg-python==0.2.0
scipy==1.13.1
numpy==1.26.4
scikit-learn==1.5.0

# Image processing (for avatars)
Pillow==10.4.0

# Data visualization
matplotlib==3.8.4

# Database
psycopg2-binary==2.9.9

# Environment
python-dotenv==1.0.1

# Development and testing
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.6.1

# Production
gunicorn==23.0.0
whitenoise==6.8.1

# Monitoring
sentry-sdk==2.17.0
